dev-requirements.txt
# Checks style, syntax, and other useful errors.
flake8

# Check basic code spelling
codespell

# Static type checking
mypy>=1.2

# Automatic code formatting
black>=22.10
isort>=5.10

# Running tests
pytest
pytest-sphinx

# Allows generation of coverage reports with pytest.
pytest-cov

# Needed for packaging and uploading to PyPi
twine>=1.11.0
setuptools
wheel

# Building docs
Sphinx>=6.0,<8

# Sphinx theme: https://sphinx-themes.org/sample-sites/furo/
furo>=2022.12.7

# Lets Sphinx parse markdown files in addition to rst.
myst-parser>=2.0

# Adds a copy button to code examples in the docs.
sphinx-copybutton>=0.5.0

# Live rebuilding and reloading of docs for developing locally.
sphinx-autobuild>=2021.3.14

# Automatically adds types to docs
sphinx-autodoc-typehints

# For documenting pydantic modules, classes and fields
autodoc_pydantic>2

# For parsing and comparing version numbers.
packaging
